:bug: Fix build plugin script

+++ bin/build-plugin.sh
@@ -1,16 +1,18 @@
 #!/bin/bash
 
 printf "Plugin name: "
-read NAME
+read -r NAME
 
 printf "Destination folder: "
-read FOLDER
+read -r FOLDER
 
 printf "Include Grunt support (Y/n): "
-read GRUNT
+read -r GRUNT
 
-printf "Initialise new git repo (y/n): "
-read NEWREPO
+printf "Initialise new git repo (y/N): "
+read -r NEWREPO
+
+set -e
 
 DEFAULT_NAME="WP Plugin Template"
 DEFAULT_CLASS=${DEFAULT_NAME// /_}
@@ -21,14 +23,21 @@ CLASS=${NAME// /_}
 TOKEN=$( tr '[A-Z]' '[a-z]' <<< "$CLASS")
 SLUG=${TOKEN//_/-}
 
-git clone "https://github.com/Monogramm/$DEFAULT_SLUG.git" "$FOLDER/$SLUG"
+if [ -n "$FOLDER" ]; then
+	if [ ! -d "$FOLDER/$SLUG" ]; then
+		mkdir -p "$FOLDER"
 
-echo "Removing git files..."
+		git clone "https://github.com/Monogramm/$DEFAULT_SLUG.git" "$FOLDER/$SLUG"
+	fi
 
-mkdir -p "$FOLDER"
-cd "$FOLDER/$SLUG"
+	cd "$FOLDER/$SLUG"
+fi
 
-rm -rf .git
+if [ "$NEWREPO" == "y" ]; then
+	echo "Removing git files..."
+	rm -rf .git
+fi
+
 rm bin/build-plugin.sh
 
 if [ "$GRUNT" == "n" ]; then
@@ -38,39 +47,39 @@ fi
 
 echo "Updating plugin files..."
 
-mv "$DEFAULT_SLUG.php" "$SLUG.php"
+mv "${DEFAULT_SLUG}.php" "${SLUG}.php"
 
-mv "lang/$DEFAULT_SLUG.pot" "lang/$SLUG.pot"
+mv "lang/${DEFAULT_SLUG}.pot" "lang/${SLUG}.pot"
 
-mv "includes/class-$DEFAULT_SLUG.php" "includes/class-$SLUG.php"
-mv "includes/class-$DEFAULT_SLUG-settings.php" "includes/class-$SLUG-settings.php"
+mv "includes/class-${DEFAULT_SLUG}.php" "includes/class-${SLUG}.php"
+mv "includes/class-${DEFAULT_SLUG}-settings.php" "includes/class-${SLUG}-settings.php"
 
-mv "includes/lib/class-$DEFAULT_SLUG-post-type.php" "includes/lib/class-$SLUG-post-type.php"
-mv "includes/lib/class-$DEFAULT_SLUG-taxonomy.php" "includes/lib/class-$SLUG-taxonomy.php"
-mv "includes/lib/class-$DEFAULT_SLUG-admin-api.php" "includes/lib/class-$SLUG-admin-api.php"
+mv "includes/lib/class-${DEFAULT_SLUG}-post-type.php" "includes/lib/class-${SLUG}-post-type.php"
+mv "includes/lib/class-${DEFAULT_SLUG}-taxonomy.php" "includes/lib/class-${SLUG}-taxonomy.php"
+mv "includes/lib/class-${DEFAULT_SLUG}-admin-api.php" "includes/lib/class-${SLUG}-admin-api.php"
 
-mv "tests/$DEFAULT_CLASS.php" "tests/$CLASS.php"
+mv "tests/${DEFAULT_CLASS}_Test.php" "tests/${CLASS}.php"
 
-for f in README.md readme.txt 'uninstall.php' "$SLUG.php" admin/js/*.js "lang/$SLUG.pot" "includes/class-$SLUG.php" "includes/class-$SLUG-settings.php" "includes/lib/class-$SLUG-post-type.php" "includes/lib/class-$SLUG-taxonomy.php" "includes/lib/class-$SLUG-admin-api.php" "tests/$CLASS.php" '.env' '.travis.yml' '.phpcs.xml.dist' '.github/PULL_REQUEST_TEMPLATE.md' '.gitlab/merge_request_templates/merge_request_template.md'
+for f in README.md readme.txt 'uninstall.php' "$SLUG.php" assets/js/*.js "lang/$SLUG.pot" "includes/class-$SLUG.php" "includes/class-$SLUG-settings.php" "includes/lib/class-$SLUG-post-type.php" "includes/lib/class-$SLUG-taxonomy.php" "includes/lib/class-$SLUG-admin-api.php" "tests/$CLASS.php" '.env' '.travis.yml' '.phpcs.xml.dist' '.github/PULL_REQUEST_TEMPLATE.md' '.gitlab/merge_request_templates/merge_request_template.md'
 do
 	cp "$f" "$f.tmp"
-	sed "s/$DEFAULT_NAME/$NAME/g" "$f.tmp" > "$f"
+	sed "s/${DEFAULT_NAME}/${NAME}/g" "$f.tmp" > "$f"
 	rm "$f.tmp"
 
 	cp "$f" "$f.tmp"
-	sed "s/$DEFAULT_SLUG/$SLUG/g" "$f.tmp" > "$f"
+	sed "s/${DEFAULT_SLUG}/${SLUG}/g" "$f.tmp" > "$f"
 	rm "$f.tmp"
 
 	cp "$f" "$f.tmp"
-	sed "s/$DEFAULT_TOKEN/$TOKEN/g" "$f.tmp" > "$f"
+	sed "s/${DEFAULT_TOKEN}/${TOKEN}/g" "$f.tmp" > "$f"
 	rm "$f.tmp"
 
 	cp "$f" "$f.tmp"
-	sed "s/$DEFAULT_CLASS/$CLASS/g" "$f.tmp" > "$f"
+	sed "s/${DEFAULT_CLASS}/${CLASS}/g" "$f.tmp" > "$f"
 	rm "$f.tmp"
 done
 
-if [ "$NEWREPO" == "y" ]; then
+if [ "${NEWREPO}" == "y" ]; then
 	echo "Initialising new git repo..."
 	git init
 fi
